out_logdna: enhance the plugin to mimic native logdna-agent event format
This commit adds logic to pack 'label' and 'annotation' if they are detected
to LogDNA request payload. It enables better handling of Kubernetes metadata
when seding logs to LogDNA.

Added a new function `retrieve_log_entry_for_key` to verify if a configuration
key exists in an event and retrieve its log entry. This function checks for
the presence of a specific key in the log event and returns the associated
log entry if found. It enables transmitting only the log record associated
to key to LogDNA.

Modified the User-Agent header value to mimic the native `logdna-agent` format.
This change enhances parsing compatibility with LogDNA.

Signed-off-by: Mirko Lazarevic <lazarevic.mirko@gmail.com>